separated stores and configs
